First working linux version in 64 bits for ZC 2.50.1 There are still many things to fix/clear as multiple warnings that appear due the bit size change.

+++ ffasm.cpp
@@ -948,7 +948,7 @@ int parse_script_file(ffscript **script, const char *path, bool report_success)
 
 int set_argument(char *argbuf, ffscript **script, int com, int argument)
 {
-    long *arg;
+    int *arg;
     
     if(argument)
     {
